Decorations refactoring
