+++ app/gui/src/project-view/components/MarkdownEditor/codemirror/decoration/table.ts
@@ -1,10 +1,10 @@
+import { VueDecorationWidget } from '@/components/MarkdownEditor/codemirror/decoration/VueDecorationWidget'
 import { nodeRange } from '@/components/MarkdownEditor/markdown/trees'
 import TableEditor from '@/components/MarkdownEditor/TableEditor.vue'
 import { type VueHost } from '@/components/VueHostRender.vue'
 import type { Text } from '@codemirror/state'
-import { Decoration, WidgetType } from '@codemirror/view'
+import { Decoration } from '@codemirror/view'
 import type { SyntaxNode, SyntaxNodeRef } from '@lezer/common'
-import { h, markRaw } from 'vue'
 import { Range } from 'ydoc-shared/util/data/range'
 
 /** Extension that uses a Vue component CodeMirror widget to render Markdown tables. */
@@ -31,41 +31,8 @@ export function decorateTable(
   }
 }
 
-class TableWidget extends WidgetType {
-  private container: HTMLElement | undefined
-  private vueHostRegistration: { unregister: () => void } | undefined
-
-  constructor(
-    private readonly props: { source: Text; parsed: SyntaxNode },
-    private readonly vueHost: VueHost,
-  ) {
-    super()
-  }
-
-  override get estimatedHeight() {
-    return -1
-  }
-
-  override toDOM(): HTMLElement {
-    if (!this.container) {
-      const container = markRaw(document.createElement('div'))
-      container.className = 'cm-table-editor'
-      this.vueHostRegistration = this.vueHost.register(
-        () =>
-          h(TableEditor, {
-            source: this.props.source,
-            parsed: this.props.parsed,
-            onEdit: () => console.log('onEdit'),
-          }),
-        container,
-      )
-      this.container = container
-    }
-    return this.container
-  }
-
-  override destroy() {
-    this.vueHostRegistration?.unregister()
-    this.container = undefined
+class TableWidget extends VueDecorationWidget<{ source: Text; parsed: SyntaxNode }> {
+  constructor(props: { source: Text; parsed: SyntaxNode }, vueHost: VueHost) {
+    super(TableEditor, props, vueHost, 'cm-table-editor')
   }
 }
